Added callback context for connect, disconnect, and BLEUART.
This allows a pointer to an object to be passed in while registering for callbacks. In turn, when the callback is invoked, the pointer will be included. This allows an instance added information, including an instance handler, for the more C++ types. This pattern is similar to the pattern followed in POSIX for thread creation.